digitization-pipeline/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
celery==5.3.4
redis==5.0.1
boto3==1.34.0
google-cloud-vision==3.4.5
langdetect==1.0.9
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0